WebKit::isTestingIPC is used from IPC::, making IPC:: depend on WebKit:: https://bugs.webkit.org/show_bug.cgi?id=243778 rdar://problem/98441549
Reviewed by Antti Koivisto.

Move WebKit::isTestingIPC implementation to IPC:: in IPCUtilities.h

* Source/WebKit/GPUProcess/graphics/RemoteGraphicsContextGLCocoa.cpp:
* Source/WebKit/GPUProcess/webrtc/LibWebRTCCodecsProxy.mm:
(WebKit::LibWebRTCCodecsProxy::createH264Decoder):
(WebKit::LibWebRTCCodecsProxy::createH265Decoder):
(WebKit::LibWebRTCCodecsProxy::createVP9Decoder):
(WebKit::LibWebRTCCodecsProxy::createEncoder):
* Source/WebKit/GPUProcess/webrtc/RemoteSampleBufferDisplayLayer.cpp:
* Source/WebKit/Platform/IPC/IPCUtilities.cpp: Added.
(IPC::isTestingIPC):
(IPC::startTestingIPC):
(IPC::stopTestingIPC):
* Source/WebKit/Platform/IPC/IPCUtilities.h: Added.
(IPC::isTestingIPC):
* Source/WebKit/Platform/IPC/cocoa/ConnectionCocoa.mm:
* Source/WebKit/Shared/IPCConnectionTester.cpp:
* Source/WebKit/Shared/IPCStreamTester.cpp:
* Source/WebKit/Shared/IPCTester.cpp:
(WebKit::IPCTester::startMessageTesting):
(WebKit::IPCTester::createStreamTester):
(WebKit::IPCTester::createConnectionTester):
(WebKit::isTestingIPC): Deleted.
* Source/WebKit/Shared/IPCTester.h:
(WebKit::isTestingIPC): Deleted.
* Source/WebKit/Shared/ThreadSafeObjectHeap.h:
* Source/WebKit/Sources.txt:
* Source/WebKit/WebKit.xcodeproj/project.pbxproj:

Canonical link: https://commits.webkit.org/253421@main
