+++ Source/WebKit/Sources.txt
@@ -198,6 +198,7 @@ Platform/IPC/DaemonDecoder.cpp
 Platform/IPC/DaemonEncoder.cpp
 Platform/IPC/Decoder.cpp
 Platform/IPC/Encoder.cpp
+Platform/IPC/IPCUtilities.cpp
 Platform/IPC/JSIPCBinding.cpp
 Platform/IPC/MessageReceiveQueueMap.cpp
 Platform/IPC/MessageReceiverMap.cpp
